Улучшает подключение к серверу
