+++ js/global-constant.js
@@ -1,0 +1,9 @@
+const form = document.querySelector('.img-upload__form');
+const hashtag = form.querySelector('.text__hashtags');
+const description = form.querySelector('.text__description');
+const uploadFile = document.querySelector('#upload-file');
+const modal = document.querySelector('.img-upload__overlay');
+
+const buttonSubmit = document.querySelector('.img-upload__submit');
+
+export {form, hashtag, description, uploadFile, modal, buttonSubmit};
