+++ js/main.js
@@ -4,4 +4,20 @@ import './big-picture.js';
 import './user-form.js'; // Почему работает без импорта?
 import './validate-form.js';
 import './editing-picture.js';
-import './api.js';
+// import { getData, sendData } from './api.js';
+
+
+// const setHui = async function (data) {
+//   try {
+//     await sendData(data);
+//   } catch {
+//     showErrorMessage();
+//   }
+// };
+
+// try {
+//   const data = await getData();
+//   renderGallery(data);
+// } catch (err) {
+//   showAllert(err.message);
+// }
