+++ js/user-form.js
@@ -1,12 +1,13 @@
-import {ZOOM_INITIAL, ADD_LISTENER_SCALES_CONTROL, REMOVE_LISTENER_SCALES_CONTROL, INIT_LEVEL_SLIDER, ADD_LISTENER_EFFECTS, REMOVE_LISTENER_EFFECTS} from './editing-picture.js';
-import { validateForm } from './validate-form.js';
-
-const form = document.querySelector('.img-upload__form');
-const hashtag = form.querySelector('.text__hashtags');
-const description = form.querySelector('.text__description');
-const modal = document.querySelector('.img-upload__overlay');
-const uploadFile = document.querySelector('#upload-file');
+import { form, hashtag, description, uploadFile, modal } from './global-constant.js';
+import { ZOOM_INITIAL, ADD_LISTENER_SCALES_CONTROL, REMOVE_LISTENER_SCALES_CONTROL, INIT_LEVEL_SLIDER, ADD_LISTENER_EFFECTS, REMOVE_LISTENER_EFFECTS } from './editing-picture.js';
+import { removeListenerModal, addListenerModal } from './form-listener-submit.js';
 const buttonClose = form.querySelector('#upload-cancel');
+
+const cleanForm = function () {
+  uploadFile.value = '';
+  hashtag.value = '';
+  description.value = '';
+};
 
 const onFormKeydown = function (evt) {
   if (evt.key === 'Escape' && document.activeElement !== description && document.activeElement !== hashtag && !document.contains(document.querySelector('.error')) && !document.contains(document.querySelector('.success'))) {
@@ -14,10 +15,8 @@ const onFormKeydown = function (evt) {
     document.body.classList.remove('modal-open');
     modal.classList.add('hidden');
     document.addEventListener('keydown', onFormKeydown);
-    form.removeEventListener('submit', validateForm);
-    uploadFile.value = '';
-    hashtag.value = '';
-    description.value = '';
+    removeListenerModal();
+    cleanForm();
 
     if (document.querySelector('.img-upload__field-wrapper').contains(document.querySelector('.pristine-error'))) {
       document.querySelector('.pristine-error').remove();
@@ -36,7 +35,7 @@ const openModal = function () {
   modal.classList.remove('hidden');
   document.body.classList.add('modal-open');
   document.addEventListener('keydown', onFormKeydown);
-  form.addEventListener('submit', validateForm);
+  addListenerModal();
 
   ADD_LISTENER_SCALES_CONTROL();
   ADD_LISTENER_EFFECTS();
@@ -48,11 +47,9 @@ const openModal = function () {
 const closeModal = function () {
   document.body.classList.remove('modal-open');
   modal.classList.add('hidden');
-  uploadFile.value = '';
-  hashtag.value = '';
-  description.value = '';
+  cleanForm();
   document.removeEventListener('keydown', onFormKeydown);
-  form.removeEventListener('submit', validateForm);
+  removeListenerModal();
 
   if (document.querySelector('.img-upload__field-wrapper').contains(document.querySelector('.pristine-error'))) {
     document.querySelector('.pristine-error').remove();
@@ -75,4 +72,4 @@ buttonClose.addEventListener('click', () => {
   closeModal();
 });
 
-export {closeModal};
+export { closeModal };
